requirements.txt
# required python >= 3.9
python-telegram-bot==13.5
requests==2.25.1
nltk==3.6.3
python-dotenv==0.19.2
google-generativeai==0.5.4
oslo.config==9.4.0
oslo.utils==7.1.0